Rearchitect USB probing.
Instead of setting up dummy operations, fully initialize a Hid that
we're probing. This has two benefits: we reduce startup time if we guess
right on the first try, and it makes it much easier to have complex
probing logic inside each implementation, instead of dumping it all in
one giant function.