Avoid stack overflow.
Commit a03b14bc is causing us to overflow the system workqueue stack
when we submit a request to redraw the display when locking as a
response to a hid GetRequest, which gets handled on the system
workqueue instead of the deferred write workqueue.

Optimize the stack usage of SOCD parsing, and while we're at it, bump
the system workqueue stack size a bit.